+++ gilded_rose.rb
@@ -1,64 +1,74 @@
+require 'itm/factory'
+require 'itm/quality_updater'
+require 'itm/quality_updater/generic'
+require 'itm/quality_updater/aged_brie'
+require 'itm/quality_updater/backstage_pass'
+require 'itm/quality_updater/sulfuras'
+require 'itm/quality_updater/default'
+
+
 def update_quality(items)
-  items.each do |item|
-    if item.name != 'Aged Brie' && item.name != 'Backstage passes to a TAFKAL80ETC concert'
-
-      if item.quality > 0
-        if item.name != 'Sulfuras, Hand of Ragnaros'
-          item.quality -= 1
-        end
-      end
-
-
-    else
-      if item.quality < 50
-        item.quality += 1
-        if item.name == 'Backstage passes to a TAFKAL80ETC concert'
-
-          if item.sell_in < 11
-            if item.quality < 50
-              item.quality += 1
-            end
-          end
-
-
-          if item.sell_in < 6
-            if item.quality < 50
-              item.quality += 1
-            end
-          end
-
-        end
-      end
-    end
-
-    if item.name != 'Sulfuras, Hand of Ragnaros'
-      item.sell_in -= 1
-    end
-
-    if item.sell_in < 0
-      if item.name != "Aged Brie"
-
-        if item.name != 'Backstage passes to a TAFKAL80ETC concert'
-
-          if item.quality > 0
-            if item.name != 'Sulfuras, Hand of Ragnaros'
-              item.quality -= 1
-            end
-          end
-
-        else
-          item.quality = item.quality - item.quality
-        end
-
-      else
-
-        if item.quality < 50
-          item.quality += 1
-        end
-
-      end
-    end
-  end
+  items.map { |i| Itm::QualityUpdater.new.call(i) }
+#  items.each do |item|
+#    if item.name != 'Aged Brie' && item.name != 'Backstage passes to a TAFKAL80ETC concert'
+#
+#      if item.quality > 0
+#        if item.name != 'Sulfuras, Hand of Ragnaros'
+#          item.quality -= 1
+#        end
+#      end
+#
+#
+#    else
+#      if item.quality < 50
+#        item.quality += 1
+#        if item.name == 'Backstage passes to a TAFKAL80ETC concert'
+#
+#          if item.sell_in < 11
+#            if item.quality < 50
+#              item.quality += 1
+#            end
+#          end
+#
+#
+#          if item.sell_in < 6
+#            if item.quality < 50
+#              item.quality += 1
+#            end
+#          end
+#
+#        end
+#      end
+#    end
+#
+#    if item.name != 'Sulfuras, Hand of Ragnaros'
+#      item.sell_in -= 1
+#    end
+#
+#    if item.sell_in < 0
+#      if item.name != "Aged Brie"
+#
+#        if item.name != 'Backstage passes to a TAFKAL80ETC concert'
+#
+#          if item.quality > 0
+#            if item.name != 'Sulfuras, Hand of Ragnaros'
+#              item.quality -= 1
+#            end
+#          end
+#
+#        else
+#          item.quality = item.quality - item.quality
+#        end
+#
+#      else
+#
+#        if item.quality < 50
+#          item.quality += 1
+#        end
+#
+#      end
+#    end
+#  end
 end
 
 # DO NOT CHANGE THINGS BELOW -----------------------------------------
