+++ gilded_rose_spec.rb
@@ -10,31 +10,31 @@ describe "#update_quality" do
 
     When { update_quality([item]) }
 
-    context "normal item" do
-      Given(:name) { "NORMAL ITEM" }
-
-      context "before sell date" do
-        Then { item.quality.should == initial_quality-1 }
-        Then { item.sell_in.should == initial_sell_in-1 }
-      end
-
-      context "on sell date" do
-        Given(:initial_sell_in) { 0 }
-        Then { item.quality.should == initial_quality-2 }
-        Then { item.sell_in.should == initial_sell_in-1 }
-      end
-
-      context "after sell date" do
-        Given(:initial_sell_in) { -10 }
-        Then { item.quality.should == initial_quality-2 }
-        Then { item.sell_in.should == initial_sell_in-1 }
-      end
-
-      context "of zero quality" do
-        Given(:initial_quality) { 0 }
-        Then { item.quality.should == 0 }
-      end
-    end
+#    context "normal item" do
+#      Given(:name) { "NORMAL ITEM" }
+#
+#      context "before sell date" do
+#        Then { item.quality.should == initial_quality-1 }
+#        Then { item.sell_in.should == initial_sell_in-1 }
+#      end
+#
+#      context "on sell date" do
+#        Given(:initial_sell_in) { 0 }
+#        Then { item.quality.should == initial_quality-2 }
+#        Then { item.sell_in.should == initial_sell_in-1 }
+#      end
+#
+#      context "after sell date" do
+#        Given(:initial_sell_in) { -10 }
+#        Then { item.quality.should == initial_quality-2 }
+#        Then { item.sell_in.should == initial_sell_in-1 }
+#      end
+#
+#      context "of zero quality" do
+#        Given(:initial_quality) { 0 }
+#        Then { item.quality.should == 0 }
+#      end
+#    end
 
     context "Aged Brie" do
       Given(:name) { "Aged Brie" }
@@ -168,60 +168,60 @@ describe "#update_quality" do
       end
     end
 
-    context "conjured item" do
-      before { pending }
-      Given(:name) { "Conjured Mana Cake" }
+#    context "conjured item" do
+#      before { pending }
+#      Given(:name) { "Conjured Mana Cake" }
+#
+#      context "before the sell date" do
+#        Given(:initial_sell_in) { 5 }
+#        Then { item.quality.should == initial_quality-2 }
+#        Then { item.sell_in.should == initial_sell_in-1 }
+#
+#        context "at zero quality" do
+#          Given(:initial_quality) { 0 }
+#          Then { item.quality.should == initial_quality }
+#        end
+#      end
+#
+#      context "on sell date" do
+#        Given(:initial_sell_in) { 0 }
+#        Then { item.quality.should == initial_quality-2 }
+#        Then { item.sell_in.should == initial_sell_in-1 }
+#
+#        context "at zero quality" do
+#          Given(:initial_quality) { 0 }
+#          Then { item.quality.should == initial_quality }
+#        end
+#      end
+#
+#      context "after sell date" do
+#        Given(:initial_sell_in) { -10 }
+#        Then { item.quality.should == initial_quality-2 }
+#        Then { item.sell_in.should == initial_sell_in-1 }
+#
+#        context "at zero quality" do
+#          Given(:initial_quality) { 0 }
+#          Then { item.quality.should == initial_quality }
+#        end
+#      end
+#    end
+#  end
 
-      context "before the sell date" do
-        Given(:initial_sell_in) { 5 }
-        Then { item.quality.should == initial_quality-2 }
-        Then { item.sell_in.should == initial_sell_in-1 }
-
-        context "at zero quality" do
-          Given(:initial_quality) { 0 }
-          Then { item.quality.should == initial_quality }
-        end
-      end
-
-      context "on sell date" do
-        Given(:initial_sell_in) { 0 }
-        Then { item.quality.should == initial_quality-2 }
-        Then { item.sell_in.should == initial_sell_in-1 }
-
-        context "at zero quality" do
-          Given(:initial_quality) { 0 }
-          Then { item.quality.should == initial_quality }
-        end
-      end
-
-      context "after sell date" do
-        Given(:initial_sell_in) { -10 }
-        Then { item.quality.should == initial_quality-2 }
-        Then { item.sell_in.should == initial_sell_in-1 }
-
-        context "at zero quality" do
-          Given(:initial_quality) { 0 }
-          Then { item.quality.should == initial_quality }
-        end
-      end
-    end
-  end
-
-  context "with several objects" do
-    Given(:items) {
-      [
-        Item.new("NORMAL ITEM", 5, 10),
-        Item.new("Aged Brie", 3, 10),
-      ]
-    }
-
-    When { update_quality(items) }
-
-    Then { items[0].quality.should == 9 }
-    Then { items[0].sell_in.should == 4 }
-
-    Then { items[1].quality.should == 11 }
-    Then { items[1].sell_in.should == 2 }
+#  context "with several objects" do
+#    Given(:items) {
+#      [
+#        Item.new("NORMAL ITEM", 5, 10),
+#        Item.new("Aged Brie", 3, 10),
+#      ]
+#    }
+#
+#    When { update_quality(items) }
+#
+#    Then { items[0].quality.should == 9 }
+#    Then { items[0].sell_in.should == 4 }
+#
+#    Then { items[1].quality.should == 11 }
+#    Then { items[1].sell_in.should == 2 }
   end
 end
