+++ itm/factory.rb
@@ -1,11 +1,11 @@
-class Item
+class Itm
   class Factory
     def call(name)
       case name
       when /Aged Brie/
         QualityUpdater::AgedBrie.new
       when /Backstage passes to a TAFKAL80ETC concert/
-        QualityUpdater::BackstageConcert.new
+        QualityUpdater::BackstagePass.new
       when /Sulfuras, Hand of Ragnaros/
         QualityUpdater::Sulfuras.new
       else
