+++ itm/quality_updater/backstage_concert.rb
@@ -1,7 +1,7 @@
-class Item
+class Itm
   class Factory
     class QualityUpdater
-      class Sulfuras
+      class BackstagePass < Generic
       end
     end
   end
