+++ itm/quality_updater/default.rb
@@ -1,12 +1,9 @@
-class Item
+class Itm
   class Factory
     class QualityUpdater
-      class Generic
-
-        UpdateNotImplemented = Class.new(NotImplementedError)
+      class Default
 
         def update(item)
-          fail UpdateNotImplemented
         end
 
 
@@ -23,3 +20,4 @@ class Item
     end
   end
 end
+
