+++ itm/quality_updater/sulfuras.rb
@@ -1,0 +1,13 @@
+class Itm
+  class Factory
+    class QualityUpdater
+      class Sulfuras < Generic
+
+        def update(item)
+
+        end
+      end
+    end
+  end
+end
+
